Fixed some gcc warnings, deleted trailing whitespaces

+++ cnpy.cpp
@@ -57,15 +57,15 @@ template<> std::vector<char>& cnpy::operator+=(std::vector<char>& lhs, const cha
     return lhs;
 }
 
-void cnpy::parse_npy_header(FILE* fp, unsigned int& word_size, unsigned int*& shape, unsigned int& ndims, bool& fortran_order) {  
+void cnpy::parse_npy_header(FILE* fp, unsigned int& word_size, unsigned int*& shape, unsigned int& ndims, bool& fortran_order) {
     char buffer[256];
-    size_t res = fread(buffer,sizeof(char),11,fp);       
+    size_t res = fread(buffer,sizeof(char),11,fp);
     if(res != 11)
         throw std::runtime_error("parse_npy_header: failed fread");
     std::string header = fgets(buffer,256,fp);
     assert(header[header.size()-1] == '\n');
 
-    int loc1, loc2;
+    size_t loc1, loc2;
 
     //fortran order
     loc1 = header.find("fortran_order")+16;
@@ -76,7 +76,7 @@ void cnpy::parse_npy_header(FILE* fp, unsigned int& word_size, unsigned int*& sh
     loc2 = header.find(")");
     std::string str_shape = header.substr(loc1+1,loc2-loc1-1);
     if(str_shape[str_shape.size()-1] == ',') ndims = 1;
-    else ndims = std::count(str_shape.begin(),str_shape.end(),',')+1;
+    else ndims = (unsigned)std::count(str_shape.begin(),str_shape.end(),',')+1;
     shape = new unsigned int[ndims];
     for(unsigned int i = 0;i < ndims;i++) {
         loc1 = str_shape.find(",");
@@ -85,7 +85,7 @@ void cnpy::parse_npy_header(FILE* fp, unsigned int& word_size, unsigned int*& sh
     }
 
     //endian, word size, data type
-    //byte order code | stands for not applicable. 
+    //byte order code | stands for not applicable.
     //not sure when this applies except for byte array
     loc1 = header.find("descr")+9;
     bool littleEndian = (header[loc1] == '<' || header[loc1] == '|' ? true : false);
@@ -134,7 +134,7 @@ cnpy::NpyArray load_the_npy_file(FILE* fp) {
     arr.word_size = word_size;
     arr.shape = std::vector<unsigned int>(shape,shape+ndims);
     delete[] shape;
-    arr.data = new char[size*word_size];    
+    arr.data = new char[size*word_size];
     arr.fortran_order = fortran_order;
     size_t nread = fread(arr.data,word_size,size,fp);
     if(nread != size)
@@ -148,7 +148,7 @@ cnpy::npz_t cnpy::npz_load(std::string fname) {
     if(!fp) printf("npz_load: Error! Unable to open file %s!\n",fname.c_str());
     assert(fp);
 
-    cnpy::npz_t arrays;  
+    cnpy::npz_t arrays;
 
     while(1) {
         std::vector<char> local_header(30);
@@ -166,7 +166,7 @@ cnpy::npz_t cnpy::npz_load(std::string fname) {
         if(vname_res != name_len)
             throw std::runtime_error("npz_load: failed fread");
 
-        //erase the lagging .npy        
+        //erase the lagging .npy
         varname.erase(varname.end()-4,varname.end());
 
         //read in the extra field
@@ -182,7 +182,7 @@ cnpy::npz_t cnpy::npz_load(std::string fname) {
     }
 
     fclose(fp);
-    return arrays;  
+    return arrays;
 }
 
 cnpy::NpyArray cnpy::npz_load(std::string fname, std::string varname) {
@@ -191,7 +191,7 @@ cnpy::NpyArray cnpy::npz_load(std::string fname, std::string varname) {
     if(!fp) {
         printf("npz_load: Error! Unable to open file %s!\n",fname.c_str());
         abort();
-    }       
+    }
 
     while(1) {
         std::vector<char> local_header(30);
@@ -205,7 +205,7 @@ cnpy::NpyArray cnpy::npz_load(std::string fname, std::string varname) {
         //read in the variable name
         unsigned short name_len = *(unsigned short*) &local_header[26];
         std::string vname(name_len,' ');
-        size_t vname_res = fread(&vname[0],sizeof(char),name_len,fp);      
+        size_t vname_res = fread(&vname[0],sizeof(char),name_len,fp);
         if(vname_res != name_len)
             throw std::runtime_error("npz_load: failed fread");
         vname.erase(vname.end()-4,vname.end()); //erase the lagging .npy
@@ -237,7 +237,7 @@ cnpy::NpyArray cnpy::npy_load(std::string fname) {
 
     if(!fp) {
         printf("npy_load: Error! Unable to open file %s!\n",fname.c_str());
-        abort();  
+        abort();
     }
 
     NpyArray arr = load_the_npy_file(fp);
@@ -245,6 +245,3 @@ cnpy::NpyArray cnpy::npy_load(std::string fname) {
     fclose(fp);
     return arr;
 }
-
-
-
